move FlatRender::op() to .cpp file
