+++ include/render.h
@@ -21,9 +21,7 @@ struct OnOffRenderer : public Renderer {
   Color color;
   OnOffRenderer(World world, Color background_color = BLACK,
                 Color _color = WHITE)
-      : Renderer(world, background_color) {
-    color = _color;
-  };
+      : Renderer(world, background_color), color{_color} {};
 
   Color operator()(Ray ray) {
     return world.ray_intersection(ray).hit ? color : background_color;
@@ -37,13 +35,4 @@ struct FlatRenderer : public Renderer {
   Color operator()(Ray ray);
 };
 
-Color FlatRenderer::operator()(Ray ray) {
-  HitRecord intersection = world.ray_intersection(ray);
-  if (!intersection.hit)
-    return background_color;
-
-  Material material = intersection.shape->material;
-  return (material.brdf->pigment->get_color(intersection.surface_point) +
-          material.emitted_radiance->get_color(intersection.surface_point));
-}
 #endif
